added v1.2
I swapped the 4 smaller thumb keys and the bottom of the middle stack. Also swapped the up and down arrow key to be more intuitive to my brain. The swapping around on the thumb clusters means that the copy and paste shortcuts can be used more easily alongside the mouse, without having to lift up my right hand from the mouse to copy/paste text. Swapped the lock and screenshot buttons for the same reason and I found I was accidentally triggering the lock key quite a bit so hopefully this swap will resolve that.